+++ Sources/SuperwallKit/Network/Device Helper/DeviceHelper.swift
@@ -207,22 +207,11 @@ class DeviceHelper {
     #endif
 	}
 
-  var fontScale: Double {
+  var fontSize: Int {
     #if os(visionOS)
-    return 1.0
+    return 16
     #else
-    let scaledBodyPointSize = UIFontMetrics.default.scaledValue(for: 17.0)
-    return Double(scaledBodyPointSize / 17.0)
-    #endif
-  }
-
-  var preferredContentSizeCategory: String {
-    #if os(visionOS)
-    return "unspecified"
-    #else
-    let category = UIApplication.sharedApplication?.preferredContentSizeCategory
-      ?? UIScreen.main.traitCollection.preferredContentSizeCategory
-    return category.rawValue
+    return Int(UIFontMetrics.default.scaledValue(for: 16.0).rounded())
     #endif
   }
 
@@ -625,8 +614,7 @@ class DeviceHelper {
       timezoneOffset: Int(TimeZone.current.secondsFromGMT()),
       radioType: radioType,
       interfaceStyle: interfaceStyle,
-      fontScale: fontScale,
-      preferredContentSizeCategory: preferredContentSizeCategory,
+      fontSize: fontSize,
       isLowPowerModeEnabled: isLowPowerModeEnabled == "true",
       isApplePayAvailable: true,
       bundleId: bundleId,