+++ Sources/SuperwallKit/Paywall/View Controller/Web View/Templating/Models/DeviceTemplate.swift
@@ -29,8 +29,7 @@ struct DeviceTemplate: Codable {
   var timezoneOffset: Int
   var radioType: String
   var interfaceStyle: String
-  var fontScale: Double
-  var preferredContentSizeCategory: String
+  var fontSize: Int
   var isLowPowerModeEnabled: Bool
   var isApplePayAvailable: Bool
   var bundleId: String